transcribator-services/async/requirements.txt

torch==2.6.0
torchvision==0.21.0
aio_pika==9.5.8
faster_whisper==1.2.1
fsspec==2026.1.0
s3fs==2026.1.0
librosa==0.11.0
nats_py==2.12.0
numpy==1.26.4
python-dotenv==1.2.1
scipy==1.15.3
redis==5.2.1
